Move db utils to ./server/db/
